+++ src/plugins/WorkloadCharacterisation.h
@@ -9,6 +9,9 @@
 #include "core/Plugin.h"
 
 #include <mutex>
+#include <string>
+#include <unordered_map>
+#include <vector>
 
 namespace llvm {
 class Function;
@@ -39,47 +42,47 @@ public:
   virtual void workItemClearBarrier(const WorkItem *workItem) override;
 
 private:
-  std::unordered_map<size_t, unsigned long> m_memoryOps;
+  std::unordered_map<size_t, uint32_t> m_memoryOps;
   std::unordered_map<std::string, size_t> m_computeOps;
-  std::unordered_map<unsigned, std::unordered_map<uint16_t, unsigned long>> m_branchPatterns;
-  std::unordered_map<unsigned, unsigned long> m_branchCounts;
-  std::vector<unsigned int> m_instructionsToBarrier;
-  std::unordered_map<unsigned, size_t> m_instructionWidth;
-  std::vector<unsigned int> m_instructionsPerWorkitem;
-  unsigned m_threads_invoked;
-  unsigned m_barriers_hit;
+  std::unordered_map<size_t, std::unordered_map<uint16_t, uint32_t>> m_branchPatterns;
+  std::unordered_map<size_t, uint32_t> m_branchCounts;
+  std::vector<uint32_t> m_instructionsToBarrier;
+  std::unordered_map<uint16_t, size_t> m_instructionWidth;
+  std::vector<uint32_t> m_instructionsPerWorkitem;
+  uint32_t m_threads_invoked;
+  uint32_t m_barriers_hit;
   int m_numberOfHostToDeviceCopiesBeforeKernelNamed;
   std::vector<std::string> m_hostToDeviceCopy;
   std::vector<std::string> m_deviceToHostCopy;
   std::string m_last_kernel_name;
-  std::vector<unsigned> m_instructionsBetweenLoadOrStore;
+  std::vector<uint32_t> m_instructionsBetweenLoadOrStore;
   std::unordered_map<std::string, size_t> m_loadInstructionLabels;
   std::unordered_map<std::string, size_t> m_storeInstructionLabels;
-  unsigned long m_global_memory_access;
-  unsigned long m_constant_memory_access;
-  unsigned long m_local_memory_access;
+  uint32_t m_global_memory_access;
+  uint32_t m_constant_memory_access;
+  uint32_t m_local_memory_access;
 
   struct WorkerState {
     std::unordered_map<std::string, size_t> *computeOps;
-    std::unordered_map<size_t, unsigned long> *memoryOps;
+    std::unordered_map<size_t, uint32_t> *memoryOps;
     bool previous_instruction_is_branch;
     std::string target1, target2;
-    unsigned branch_loc;
-    std::unordered_map<unsigned, std::vector<bool>> *branchOps;
-    unsigned threads_invoked;
-    unsigned barriers_hit;
-    unsigned instruction_count;
-    unsigned ops_between_load_or_store;
-    unsigned workitem_instruction_count;
-    unsigned long global_memory_access_count;
-    unsigned long constant_memory_access_count;
-    unsigned long local_memory_access_count;
-    std::vector<unsigned> *instructionsBetweenBarriers;
-    std::vector<unsigned> *instructionsPerWorkitem;
-    std::unordered_map<unsigned, size_t> *instructionWidth;
+    uint32_t branch_loc;
+    std::unordered_map<size_t, std::vector<bool>> *branchOps;
+    uint32_t threads_invoked;
+    uint32_t barriers_hit;
+    uint32_t instruction_count;
+    uint32_t ops_between_load_or_store;
+    uint32_t workitem_instruction_count;
+    uint32_t global_memory_access_count;
+    uint32_t constant_memory_access_count;
+    uint32_t local_memory_access_count;
+    std::vector<uint32_t> *instructionsBetweenBarriers;
+    std::vector<uint32_t> *instructionsPerWorkitem;
+    std::unordered_map<uint16_t, size_t> *instructionWidth;
     std::unordered_map<std::string, size_t> *loadInstructionLabels;
     std::unordered_map<std::string, size_t> *storeInstructionLabels;
-    std::vector<unsigned> *instructionsBetweenLoadOrStore;
+    std::vector<uint32_t> *instructionsBetweenLoadOrStore;
   };
   static THREAD_LOCAL WorkerState m_state;
